adding new 175 pat project removing 150 file
Old example workflow file didn't open and run. I re-made it as 175 in a
new Sample Workflows folder. There are still some directories that start
with "pat" at the top level of repo. Those pat projects are not complete
and or do not open correctly in current version of OpenStudio. They will
be updated at a later date.